% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSZ95.R
\name{PTSZ95}
\alias{PTSZ95}
\title{This function calculates PTDO when fix PTSZ=0.95.}
\usage{
PTSZ95(observed, expected, result)
}
\arguments{
\item{observed}{Observed single cells matrix with each column being the upper triangular of a single cell.}

\item{expected}{Underline true counts from simulation.}

\item{result}{Result form MCMCImpute function.}
}
\value{
A vector of PTDO and its SD when fixing PTSZ to be 0.95, and the threshold used in that case.
}
\description{
This function calculates PTDO when fix PTSZ=0.95.
}
\examples{
PTSZ95(observed=K562_T1_7k, expected=K562_1_true, result=T1_7k_res)
}
