% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_assess.R
\name{scHiC_assess}
\alias{scHiC_assess}
\title{This function analyzes both simulated and real datasets, depending on the inputs of the functions.}
\usage{
scHiC_assess(
  result,
  cell_index = 1,
  n,
  cell_type,
  dims = 2,
  perplexity = 10,
  seed = 1000,
  kmeans = TRUE,
  ncenters = 2
)
}
\arguments{
\item{result}{Output of MCMCImpute for simulated data or the organized results of real data.}

\item{cell_index}{Indicates which cell is used to draw heatmaps and scatterplot.}

\item{n}{Dimension of 2D contact matrix.}

\item{cell_type}{A vector of underlying true cluster.}

\item{dims}{The dimension of 2D matrix.}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than \eqn{3\times perplexity<nrow(X)-1}).}

\item{seed}{Random seed for generating t-SNE data.}

\item{kmeans}{Logical, whether apply K-means clustering on the t-SNE data.}

\item{ncenters}{Number of centers in K-means clustering analysis.}
}
\value{
A list of accuracy measurements and plots.
}
\description{
This function analyzes both simulated and real datasets, depending on the inputs of the functions.
}
\examples{
data("K562_1_true")
options(digits = 2)
scHiC_assess(result=K562_T1_4k_result)
}
