% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_hm.R
\name{scHiC_hm}
\alias{scHiC_hm}
\title{This function draws heatmap of HiC data so that we can visually compares the imputation results.}
\usage{
scHiC_hm(datvec, n, title = "Heatmap")
}
\arguments{
\item{datvec}{A vector of upper triangular mamtrix.}

\item{n}{Dimension of 2D matrix (i.e., the number of segments).}

\item{title}{The title of the heatmap.}
}
\value{
Heatmap of the matrix.
}
\description{
This function draws heatmap of HiC data so that we can visually compares the imputation results.
}
\examples{
.
data("K562_1_true")
scHiC_hm(K562_1_true[,1], 61, title="Expected")
}
