% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_Kmeans.R
\name{scHiC_Kmeans}
\alias{scHiC_Kmeans}
\title{This function conduct Kmeans clustering analysis on scHi-C data.}
\usage{
scHiC_Kmeans(
  data,
  centers,
  nstart = 50,
  iter.max = 200,
  seed = 1234,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE
)
}
\arguments{
\item{data}{The observed or imputed matirx, with each column being the uppertriangular of a single cell HiC matrix.}

\item{centers}{Either the number of clusters, say k, or a set of initial (distinct) cluster centres. If a number, a random set of (distinct) rows in x is chosen as the initial centres.}

\item{nstart}{If centers is a number, how many random sets should be chosen.}

\item{iter.max}{The maximum number of iterations allowed.}

\item{seed}{Random seed.}

\item{algorithm}{Character: may be abbreviated. Note that "Lloyd" and "Forgy" are alternative names for one algorithm.}

\item{trace}{Logical or integer number, currently only used in the default method ("Hartigan-Wong"): if positive (or true), tracing information on the progress of the algorithm is produced. Higher values may produce more tracing information.}
}
\value{
Kmeans clustering results.
}
\description{
This function conduct Kmeans clustering analysis on scHi-C data.
}
\examples{
data("GSE117874_chr1_wo_diag")
data("GSE117874_imp")
cluster=scHiC_Kmeans(GSE117874_chr1_wo_diag, centers=2, nstart=1, iter.max=1000, seed=1)
}
