% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_ROC.R
\name{scHiC_ROC}
\alias{scHiC_ROC}
\title{This package draws ROC (Receiver operating characteristic) curve to visually demonstrate ability 
to tell SZ from DO.}
\usage{
scHiC_ROC(observed, expected, result)
}
\arguments{
\item{observed}{Observed sngle cell with each column being the upper triangular of single cell.}

\item{expected}{Underline true count of simulated data.}

\item{result}{Result from MCMCImpute funciton.}
}
\value{
A plot of ROC curve.
}
\description{
This package draws ROC (Receiver operating characteristic) curve to visually demonstrate ability 
to tell SZ from DO.
}
\examples{
scHiC_ROC(observed=K562_T1_7k, expected=K562_1_true, result=T1_7k_res)
}
