% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_simulate.R
\name{scHiC_simulate}
\alias{scHiC_simulate}
\title{This function simulates single cells from 3D structure.}
\usage{
scHiC_simulate(
  data = str1,
  alpha_0,
  alpha_1,
  beta_l,
  beta_g,
  beta_m,
  gamma,
  eta,
  n_single
)
}
\arguments{
\item{data}{3D coordinates of single cell.}

\item{alpha_0}{Parameter that controls sequence depth of data.}

\item{alpha_1}{Parameter that controls sequence depth of data.}

\item{beta_l}{Parameter that controls effect size of covariate.}

\item{beta_g}{Parameter that controls effect size of covariate.}

\item{beta_m}{Parameter that controls effect size of covariate.}

\item{gamma}{Quantile that is used as the threshold.}

\item{eta}{Percent of structural zeros that are set to be common structural zeros among all single-cells.}

\item{n_single}{Number of single cells to be generated.}
}
\value{
A list of underline true count, SZ positions, and generated single cells.
}
\description{
This function simulates single cells from 3D structure.
}
\examples{
#Load 3d structure generated from SIMBA package
load("simba_3strs.rdata")
Set random seed
set.seed(1234)
#Generate 100 random type1 single cells
simudat <- scHiC_simulate(data=str1, alpha_0=5.6,alpha_1=-1, beta_l=0.9,beta_g=0.9,
beta_m=0.9,gamma=0.1,eta=0.8, n_single=10) 
}
