% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_tSNE.R
\name{scHiC_tSNE}
\alias{scHiC_tSNE}
\title{This function visualize scHi-C data using t-SNE (t-distributed stochastic neighbor embedding) 
and applying Kmeans clustering followed by xie et al. 2021.}
\usage{
scHiC_tSNE(
  data,
  cell_type,
  dims = 2,
  perplexity = 10,
  check_duplicates = FALSE,
  seed = 1234,
  title = NULL,
  kmeans = TRUE,
  ncenters
)
}
\arguments{
\item{data}{The observed matrix, with each column being the uppertriangular of a single cell HiC matrix.}

\item{cell_type}{A vector that indicates cell type.}

\item{dims}{Integer. Output dimentionality. Default=2.}

\item{perplexity}{Numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation).}

\item{check_duplicates}{Logical; Checks whether duplicates are present. It is best to make sure there are no duplicates present and set this option to FALSE, especially for large datasets (default: TRUE).}

\item{seed}{Random seed.}

\item{title}{Title of the plot.}

\item{ncenters}{Number of clusters in kmeans clustering.}
}
\value{
A stne visualization plot.
}
\description{
This function visualize scHi-C data using t-SNE (t-distributed stochastic neighbor embedding) 
and applying Kmeans clustering followed by xie et al. 2021.
}
\examples{
scHiC_tSNE(GSE117874_chr1_wo_diag, cell_type=c(rep("GM",14),rep("PBMC",18)), 
dims = 2,perplexity=10, seed=1000, title="Observed GSE117874", 
kmeans = TRUE, ncenters = 2)
}
