% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEVI.R
\name{SEVI}
\alias{SEVI}
\title{This function generates scatterplot of expected versus imputed.}
\usage{
SEVI(obsvec, expvec, impvec)
}
\arguments{
\item{obsvec}{A vector of observed single cell.}

\item{expvec}{A vector of expected single cell.}

\item{impvec}{A vector of imputed single cell.}
}
\value{
The scatterplot of expected versus imputed, with read dots being the observed zero pairs.
}
\description{
This function generates scatterplot of expected versus imputed.
}
\examples{
SEVI(obsvec=K562_T1_7k[,1], expvec=K562_1_true[,1], impvec=T1_7k_imp[,1] )
}
