% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOVI.R
\name{SOVI}
\alias{SOVI}
\title{This function generates scatterplot of observed versus imputed for nonzero observed counts.}
\usage{
SOVI(obsvec, impvec)
}
\arguments{
\item{obsvec}{A vector of observed single cell.}

\item{impvec}{A vector of imputed single cell.}
}
\value{
The scatterplot of observed versus imputed.
}
\description{
This function generates scatterplot of observed versus imputed for nonzero observed counts.
}
\examples{
data("GSE117874_imp")
data("GSE117874_chr1_wo_diag")
SOVI(obsvec = GSE117874_chr1_wo_diag[,1], impvec = GSE117874_imp[,1])
}
