% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHiC_assess.R
\name{scHiC_assess}
\alias{scHiC_assess}
\title{This function analyzes both simulated and real datasets, depending on the inputs of the functions.}
\usage{
scHiC_assess(
  scHiC,
  expected = NULL,
  result = NULL,
  imputed = NULL,
  cell_index = 1,
  n,
  cell_type,
  dims = 2,
  perplexity = 10,
  seed = 1000,
  kmeans = TRUE,
  ncenters = 2
)
}
\arguments{
\item{scHiC}{The observed data. It can take three types of formats. The preferred format is a single-cell 
matrix with each column being a vector of the upper triangular matrix without including the diagonal entries 
of the 2D matrix of a single-cell. Another types of formats are a list with each element being a 2D s
ingle-cell contact matrix, or a 3D (\eqn{n\times n\times k}) array that has k matrices of dimension \eqn{n\times n}. 
HiCImpute automatically transforms these two types of input into a matrix with each column being the vector of 
upper triangular matrix of a single-cell.}

\item{expected}{Underline true counts of the simulated data. For real data analysis, just set it as NULL.}

\item{result}{Output of MCMCImpute.}

\item{imputed}{The imputed data that has the same dimension as the observed data. This is needed for real 
data analysis. For simulated data, set it as NULL.}

\item{cell_index}{Indicates which cell is used to draw heatmaps and scatterplot.}

\item{n}{Dimension of 2D contact matrix.}

\item{cell_type}{A vector of underlying true cluster.}

\item{dims}{The dimension of 2D matrix.}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than \eqn{3\times perplexity<nrow(X)-1}).}

\item{seed}{Random seed for generating t-SNE data.}

\item{kmeans}{Logical, whether apply K-means clustering on the t-SNE data.}

\item{ncenters}{Number of centers in K-means clustering analysis.}
}
\value{
A list of accuracy measurements and plots.
}
\description{
This function analyzes both simulated and real datasets, depending on the inputs of the functions.
}
\examples{
data("K562_1_true")
options(digits = 2)
scHiC_assess(scHiC=K562_T1_7k, expected=K562_1_true, imputed=result$Impute_SZ=T1_7k_imp)
}
